requirements.txt
requests
beautifulsoup4
openai>=1.30.0
pinecone
fastapi
uvicorn[standard]
python-dotenv

